third_party/blink/web_tests/fast/css/pseudo-in-range-invalid-value-expected.txt

CONSOLE WARNING: The specified value "ABC" cannot be parsed, or is out of range.
Tests that invalid values are considered in-range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector("input[type=number]:in-range").id is "number1"
PASS document.querySelector("input[type=range]:in-range").id is "range1"
PASS document.querySelectorAll(":in-range").length is 2
PASS successfullyParsed is true

TEST COMPLETE

 
